+++ apps/blog/app/Header.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+
+
+const Header = () => {
+  return (
+    <div
+      style={{
+        background: "red",
+        position: "fixed",
+        height: "10dvh",
+        width: "100%",
+      }}
+    >
+      Header
+    </div>
+  );
+};
+
+export default Header;
