+++ apps/blog/app/globals.css
@@ -1,3 +1,5 @@
 body {
   margin: 0;
+  width: 100dvw;
+  height: 100dvh;
 }
