+++ apps/blog/package.json
@@ -24,6 +24,7 @@
     "@types/node": "^22",
     "@types/react": "19.0.8",
     "@types/react-dom": "19.0.3",
+    "@vanilla-extract/next-plugin": "^2.4.10",
     "eslint": "^9.20.0",
     "typescript": "5.7.3"
   }
